requirements.txt
anyascii==0.3.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==1.7.0
Babel==2.10.3
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
Deprecated==1.2.13
Django==3.2.16
django-cogwheels==0.3
django-dotenv==1.4.2
django-environ==0.9.0
django-extra-views==0.13.0
django-filter==22.1
django-haystack==3.2.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-phonenumber-field==5.2.0
django-redis==5.2.0
django-tables2==2.3.4
django-taggit==3.0.0
django-treebeard==4.5.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
factory-boy==3.2.1
Faker==15.1.1
html5lib==1.1
idna==3.4
l18n==2021.3
openpyxl==3.0.10
packaging==21.3
phonenumbers==8.12.57
Pillow==9.2.0
psycopg2-binary==2.9.4
purl==1.6
pycodestyle==2.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
redis==4.3.4
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tablib==3.2.1
telepath==0.3
toml==0.10.2
urllib3==1.26.12
wagtail==4.0.4
wagtail-metadata==4.0.2
wagtailcodeblock==1.28.0.0
wagtailmenus==3.1.3
webencodings==0.5.1
Willow==1.4.1
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
